Move EmptyAppBar and LinkCard components
