+++ lib/common/app_bar_gone.dart
