+++ lib/features/info/info_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:responsive_builder/responsive_builder.dart';
 
-import '../../ui/widgets/app_bar_gone.dart';
+import '../../common/app_bar_gone.dart';
 import '../../common/bottom_nav_bar/bottom_nav_bar.dart';
 import '../../common/grid_item.dart';
